rename to worktree-docker
