+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
-## python_template
+## worktree_docker
 
 ## [0.0.0]
