+++ README.md
@@ -1,4 +1,4 @@
-# python_template
+# worktree_docker
 A template repo for python projects that is set up using [pixi](https://pixi.sh). 
 
 This has basic setup for
@@ -16,12 +16,12 @@ This has basic setup for
 
 ## Continuous Integration Status
 
-[![Ci](https://github.com/blooop/python_template/actions/workflows/ci.yml/badge.svg?branch=main)](https://github.com/blooop/python_template/actions/workflows/ci.yml?query=branch%3Amain)
-[![Codecov](https://codecov.io/gh/blooop/python_template/branch/main/graph/badge.svg?token=Y212GW1PG6)](https://codecov.io/gh/blooop/python_template)
-[![GitHub issues](https://img.shields.io/github/issues/blooop/python_template.svg)](https://GitHub.com/blooop/python_template/issues/)
-[![GitHub pull-requests merged](https://badgen.net/github/merged-prs/blooop/python_template)](https://github.com/blooop/python_template/pulls?q=is%3Amerged)
-[![GitHub release](https://img.shields.io/github/release/blooop/python_template.svg)](https://GitHub.com/blooop/python_template/releases/)
-[![License](https://img.shields.io/github/license/blooop/python_template)](https://opensource.org/license/mit/)
+[![Ci](https://github.com/blooop/worktree_docker/actions/workflows/ci.yml/badge.svg?branch=main)](https://github.com/blooop/worktree_docker/actions/workflows/ci.yml?query=branch%3Amain)
+[![Codecov](https://codecov.io/gh/blooop/worktree_docker/branch/main/graph/badge.svg?token=Y212GW1PG6)](https://codecov.io/gh/blooop/worktree_docker)
+[![GitHub issues](https://img.shields.io/github/issues/blooop/worktree_docker.svg)](https://GitHub.com/blooop/worktree_docker/issues/)
+[![GitHub pull-requests merged](https://badgen.net/github/merged-prs/blooop/worktree_docker)](https://github.com/blooop/worktree_docker/pulls?q=is%3Amerged)
+[![GitHub release](https://img.shields.io/github/release/blooop/worktree_docker.svg)](https://GitHub.com/blooop/worktree_docker/releases/)
+[![License](https://img.shields.io/github/license/blooop/worktree_docker)](https://opensource.org/license/mit/)
 [![Python](https://img.shields.io/badge/python-3.10%20%7C%203.11%20%7C%203.12%20%7C%203.13-blue)](https://www.python.org/downloads/)
 [![Pixi Badge](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/prefix-dev/pixi/main/assets/badge/v0.json)](https://pixi.sh)
 
@@ -77,7 +77,7 @@ pixi run ci
 
 ## Legacy
 
-If you don't want to install rocker on your system but want to use vscode, you can run the `scripts/launch_vscode.sh` script to build and connect to a docker container. It will install rocker in a venv.  The docker container is dynamically generated using [rocker](https://github.com/osrf/rocker) and [deps rocker](https://github.com/blooop/deps_rocker).  [deps rocker](https://github.com/blooop/deps_rocker) looks at the python_template.deps.yaml file to install any required apt, pip or shell scripts and launches a container that vscode attaches to. 
+If you don't want to install rocker on your system but want to use vscode, you can run the `scripts/launch_vscode.sh` script to build and connect to a docker container. It will install rocker in a venv.  The docker container is dynamically generated using [rocker](https://github.com/osrf/rocker) and [deps rocker](https://github.com/blooop/deps_rocker).  [deps rocker](https://github.com/blooop/deps_rocker) looks at the worktree_docker.deps.yaml file to install any required apt, pip or shell scripts and launches a container that vscode attaches to. 
 
 ## Troubleshooting
