+++ docs/index.rst
@@ -1,9 +1,9 @@
-.. python_template documentation master file, created by
+.. worktree_docker documentation master file, created by
    sphinx-quickstart on Mon Nov 27 15:01:32 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to python_template's documentation!
+Welcome to worktree_docker's documentation!
 ===========================================
 
 .. toctree::
